fix: 暂时去掉浏览器使用worker计算sha1逻辑。 fix: 优化下载逻辑。

+++ tests/ft/LoadFile-HostingMode.test.ts
@@ -1,6 +1,7 @@
 /** @format */
 
 import assert = require('assert')
+import {existsSync, unlinkSync} from 'fs'
 
 const Config = require('./conf.js')
 const {join} = require('path')
@@ -11,20 +12,25 @@ const {getClient} = require('./token-util')
 const PATH_TYPE = 'HostingMode'
 
 describe('LoadFile', function () {
-  this.timeout(60000)
+  this.timeout(600000)
 
   describe('HostingMode', () => {
     it('success', async () => {
       const path_type = PATH_TYPE
       const {domain_id, drive_id} = Config['domains'][path_type]
 
-      let from = join(__dirname, `tmp/tmp-${domain_id}.txt`)
+      let from = join(__dirname, `tmp/tmp-${domain_id}-hosting.txt`)
+
+      let local_name = `tmp-${domain_id}-hosting-2.txt`
+      let to = join(__dirname, 'tmp', local_name)
+      if (existsSync(to)) unlinkSync(to)
 
       // mock 文件
-      execSync(`dd if=/dev/zero of=${from} bs=1024 count=10000`)
+      if (!existsSync(from)) execSync(`dd if=/dev/zero of=${from} bs=1024 count=10000`)
 
       var client = await getClient(path_type)
 
+      let task
       // 上传
       var cp = await client.uploadFile(
         from,
@@ -35,9 +41,11 @@ describe('LoadFile', function () {
         {
           parallel_upload: false,
           verbose: true, //显示详细日志
-
+          onReady(t) {
+            task = t
+          },
           onProgress(state, progress) {
-            console.log(state, progress)
+            console.log(state, progress, task.speed / 1024 / 1024 + 'M/s')
           },
           onPartComplete(cp, partInfo) {
             console.log('onPartComplete:', partInfo.part_number, '---done-------')
@@ -56,14 +64,15 @@ describe('LoadFile', function () {
       // 下载
       let fileInfo = await client.postAPI('/file/get', {drive_id, file_path})
 
-      console.log(fileInfo)
+      // console.log(fileInfo)
 
-      let local_name = `tmp-${domain_id}-2.txt`
-
-      var cp2 = await client.downloadFile(fileInfo, join(__dirname, 'tmp', local_name), {
+      var cp2 = await client.downloadFile(fileInfo, to, {
         verbose: true, //显示详细日志
+        onReady(t) {
+          task = t
+        },
         onProgress(state, progress) {
-          console.log(state, progress)
+          console.log(state, progress, task.speed / 1024 / 1024, 'MB/s')
         },
         onPartComplete(cp, partInfo) {
           console.log('onPartComplete:', partInfo.part_number, '---done-------')
@@ -75,6 +84,8 @@ describe('LoadFile', function () {
       assert(cp2.file_key == file_path)
 
       await client.postAPI('/file/delete', {drive_id, file_path})
+
+      if (existsSync(to)) unlinkSync(to)
     })
   })
 })
